Introducing compatibility with Backpack 4.1 Fluent Syntax.
For that, Field and Column base classes were altered (as well as
Arrayable); they now extend Backpack's CrudField and CrudColumn.

I also introduced a bit of testing just to make sure things are fine.
And updated code generator to reflect the fact that name() method is now
used as a constructor, just like its done in CrudField and CrudColumn,
which makes sense as it feels like better alternative to make().